Add files via upload
We have collected information about patients suffering from the same disease. The development of one of 5 different drugs affects the patients who helped the children's course.

Our task is to model which medicine is suitable for the identified patient with this disease. This problem is multi-class classification (multiclass classification) help score consolidation.